Big Update

+++ src/main/java/com/example/skyfast_2_0/dto/AirplaneDTO.java
@@ -1,19 +1,19 @@
-package com.example.skyfast_2_0.dto;
+ package com.example.skyfast_2_0.dto;
 
-import lombok.Data;
+ import lombok.Data;
 
-@Data
-public class AirplaneDTO {
-    private Integer id;
-    private String airplaneName;
-    private String manufacturer;
-    private String diagram;
-    private Integer speed;
-    private Integer totalLength;
-    private Integer wingspan;
-    private Integer height;
-    private String airplaneStatus;
-    private Integer seatCapacity;
-    private String airplaneImage;
-    private Integer airlineId;
-}
+ @Data
+ public class AirplaneDTO {
+     private Integer id;
+     private String airplaneName;
+     private String manufacturer;
+     private String diagram;
+     private Float speed;
+     private Float totalLength;
+     private Float wingspan;
+     private Float height;
+     private String airplaneStatus;
+     private Integer seatCapacity;
+     private String airplaneImage;
+     private Integer airlineId;
+ }
